refactor(generator): extract DependencyChecks (SM0010/0011)
